requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
cycler==0.10.0
idna==2.9
instaloader==4.4.3
kiwisolver==1.2.0
matplotlib==3.2.1
numpy==1.18.4
opencv-python==4.2.0.34
Pillow==7.1.2
pyparsing==2.4.7
pytesseract==0.3.4
python-dateutil==2.8.1
requests==2.23.0
scipy==1.4.1
six==1.15.0
urllib3==1.25.9